+++ DataAccess/Operations/ReservationOperations.cs
@@ -23,6 +23,7 @@ namespace DataAccess.Operations
         public Reservations DeleteReservation(int id)
         {
             var reservation = GetByPk(id);
+            if (reservation == null) { return null; }
             db.Remove(reservation);
             db.SaveChanges();
             return reservation;
