+++ DataAccess/Operations/ServiceEmployeeOperations.cs
@@ -23,6 +23,7 @@ namespace DataAccess.Operations
         public ServicesEmployees DeleteServiceemployee(int id)
         {
             var serviceEmployee = GetByPk(id);
+            if (serviceEmployee == null) { return null; }
             db.Remove(serviceEmployee);
             db.SaveChanges();
             return serviceEmployee;
